created new directory "cells" for everything cells related. Mainly cell templates, but also planning on having things like custom section classes and stims/recorders there too. Latter is subject to change
